Update FranceFSM vector info & katai struct.
Previously changes were made to the katai struct to clarify the purpose of each particular value of the transformation vector, however this obscures the significance of the structure being a transformation matrix. Instead it has been updated to use the matrix value names and adds info about the matrix itself to the description.